+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - stimulus HMR
 - fix bug : stimulus restart vite dev server when controllers.json is updated
 - split vite-plugin-symfony into 2 plugins `vite-plugin-symfony-entrypoints` and `vite-plugin-symfony-stimulus`.
+- add new tests to vite-plugin-symfony
+- doc : add mermaid charts
 
 ## v6.2.0